WEB-2039 refactor stories
